chore(lab1): restructure app
