+++ app_golang/src/templates/joke.tmpl
